requirements.txt
#add requirements(with version) to this file so that it installs easily for others

numpy==v2.3.0
pymodbus==v3.11.3
pyModbusTCP==v0.3.0
pillow==v11.3.0
python-snap7==v2.0.2
